fix conditional rendering

+++ src/components/Dashboard/StickyHeadTable.js
@@ -6,14 +6,14 @@ import UserTable from '@/components/Dashboard/UserTable'
 import { httpClient } from '@/utils/api'
 
 export default function StickyHeadTable({
-    users,
-    setUsers,
-    fetchUser,
-    permissionsList,
-    fetchLoginUserPermission,
-    edit_user_permission,
-    delete_user_permission,
-}) {
+                                            users,
+                                            setUsers,
+                                            fetchUser,
+                                            permissionsList,
+                                            fetchLoginUserPermission,
+                                            edit_user_permission,
+                                            delete_user_permission,
+                                        }) {
     const [page, setPage] = useState(0)
     const [rowsPerPage, setRowsPerPage] = useState(10)
     const [openModal, setOpenModal] = useState({ edit: false, delete: false })
@@ -55,33 +55,44 @@ export default function StickyHeadTable({
             role: userRole,
             permissions: userPermissions,
         }
+
         try {
-            const res = await httpClient.patch(
-                `/user/${currentUser.id}/`,
-                payload
-            )
+            const res = await httpClient.patch(`/user/${currentUser.id}/`, payload)
+
             alert(res.data.detail)
 
             const updatedUsers = await fetchUser()
-            setUsers(updatedUsers.users)
-            fetchLoginUserPermission()
+
+            setUsers(updatedUsers)
+            await fetchLoginUserPermission()
         } catch (err) {
-            console.log(err)
+            console.error('Error saving edit:', err)
+            alert('Failed to save the changes. Please try again.')
         }
+
         setOpenModal({ edit: false, delete: false })
     }
 
+
     const confirmDelete = async () => {
+        if (!currentUser) {
+            console.error('No user selected for deletion.')
+            return
+        }
+
         try {
             const res = await httpClient.delete(`/user/${currentUser.id}/`)
             alert(res.data.detail)
+
             const updatedUsers = await fetchUser()
-            setUsers(updatedUsers.users)
+            setUsers(updatedUsers)
         } catch (err) {
-            console.log(err)
+            console.error('Error deleting user:', err)
         }
+
         setOpenModal({ edit: false, delete: false })
     }
+
 
     return (
         <>
@@ -113,7 +124,7 @@ export default function StickyHeadTable({
                     setUserPermissions(
                         typeof e.target.value === 'string'
                             ? e.target.value.split(',')
-                            : e.target.value
+                            : e.target.value,
                     )
                 }
                 handleSave={handleSaveEdit}
